refactor: Add equipment-options-dropdown
