+++ src/app/equipment/equipment-detail/equipment-detail.component.html
@@ -4,25 +4,7 @@
     <div class="bi-device-ssd px-3">
       {{ equipment?.type?.name }}
     </div>
-    <div ngbDropdown class="d-inline-block">
-      <button type="button" class="btn btn-link bi-three-dots" id="dropdown-audit" ngbDropdownToggle>
-      </button>
-      <div ngbDropdownMenu aria-labelledby="dropdown-audit">
-        <!-- TODO: Duplicated in type-list.component.html -->
-        @if (equipment?.type?.name === 'HVAC') {
-          <h5 class="dropdown-header">
-            HVAC Specific
-          </h5>
-          <a ngbDropdownItem class="bi-link-45deg" routerLink="connect">
-            Connect Zone
-          </a>
-          <div class="dropdown-divider"></div>
-        }
-        <button ngbDropdownItem class="bi-pencil-square" (click)="rename()">
-          Rename
-        </button>
-      </div>
-    </div>
+    <app-equipment-options-dropdown [equipment]="equipment"></app-equipment-options-dropdown>
   </div>
   <div class="p-3">
     <h1>
