+++ src/app/equipment/equipment-list/equipment-list.component.ts
@@ -1,10 +1,11 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
-import {EMPTY, switchMap} from 'rxjs';
-import {AuditService} from 'src/app/shared/services/audit.service';
-import {EquipmentService} from 'src/app/shared/services/equipment.service';
-import {Equipment, EquipmentCategory} from '../../shared/model/equipment.interface';
 import {ToastService} from '@mean-stream/ngbx';
+import {EMPTY, switchMap} from 'rxjs';
+
+import {Equipment, EquipmentCategory} from '../../shared/model/equipment.interface';
+import {AuditService} from '../../shared/services/audit.service';
+import {EquipmentService} from '../../shared/services/equipment.service';
 
 @Component({
   selector: 'app-equipment-list',
@@ -64,20 +65,6 @@ export class EquipmentListComponent implements OnInit {
       zoneId: this.route.snapshot.params.zid,
       equipmentId: category.id,
     }).subscribe(res => this.auditService.currentProgress = res);
-  }
-
-  rename(item: Equipment) {
-    const kind = item.type?.name;
-    const name = prompt(`Rename ${kind}`, item.name);
-    if (!name) {
-      return;
-    }
-
-    this.equipmentService.updateEquipment({...item, name}).subscribe(({data}) => {
-      const index = this.equipments.indexOf(item);
-      this.equipments[index] = data;
-      this.toastService.success(`Rename ${kind}`, `Successfully renamed ${kind}`);
-    });
   }
 
   delete(item: Equipment) {
