+++ src/app/equipment/equipment.module.ts
@@ -1,17 +1,18 @@
-import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-
-import {EquipmentRoutingModule} from './equipment-routing.module';
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {ModalModule} from '@mean-stream/ngbx';
+import {NgbDropdownModule, NgbTooltipModule} from '@ng-bootstrap/ng-bootstrap';
+import {FormsModule as AppFormsModule} from '../shared/form/form.module';
+import {SharedModule} from '../shared/shared.module';
 import {ConnectZoneComponent} from './connect-zone/connect-zone.component';
 import {CreateEquipmentComponent} from './create-equipment/create-equipment.component';
 import {EquipmentDetailComponent} from './equipment-detail/equipment-detail.component';
-import {EquipmentMasterDetailComponent} from './equipment-master-detail/equipment-master-detail.component';
 import {EquipmentListComponent} from './equipment-list/equipment-list.component';
-import {ModalModule} from '@mean-stream/ngbx';
-import {SharedModule} from '../shared/shared.module';
-import {FormsModule as AppFormsModule} from '../shared/form/form.module';
-import {NgbDropdownModule, NgbTooltipModule} from '@ng-bootstrap/ng-bootstrap';
-import {FormsModule} from '@angular/forms';
+import {EquipmentMasterDetailComponent} from './equipment-master-detail/equipment-master-detail.component';
+import {EquipmentOptionsDropdownComponent} from './equipment-options-dropdown/equipment-options-dropdown.component';
+
+import {EquipmentRoutingModule} from './equipment-routing.module';
 
 
 @NgModule({
@@ -24,6 +25,7 @@ import {FormsModule} from '@angular/forms';
     NgbTooltipModule,
     SharedModule,
     AppFormsModule,
+    EquipmentOptionsDropdownComponent,
   ],
   declarations: [
     EquipmentListComponent,
